# [ENH] improved performance for `first/last` in `conditional_join`